optomizing common utils
